JWT实现跨域身份认证

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -9,7 +9,9 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.servlet.http.HttpServletRequest;
 
+import io.jsonwebtoken.Claims;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,12 +20,14 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import util.IdWorker;
 
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -44,6 +48,15 @@ public class UserService {
 
     @Autowired
     private RabbitTemplate rabbitTemplate;
+
+    @Autowired
+    private BCryptPasswordEncoder encoder;
+
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    @Autowired
+    private HttpServletRequest request;
 
     /**
      * 查询全部列表
@@ -98,6 +111,8 @@ public class UserService {
      */
     public void add(User user) {
         user.setId(idWorker.nextId() + "");
+        //密码加密
+        user.setPassword(encoder.encode(user.getPassword()));
         user.setFollowcount(0); //关注数
         user.setFanscount(0); //粉丝数
         user.setOnline(0L); //在线时长
@@ -123,6 +138,10 @@ public class UserService {
      * @param id
      */
     public void deleteById(String id) {
+        String token = (String) request.getAttribute("claims_admin");
+        if (token == null || "".equals(token)) {
+            throw new RuntimeException("权限不足!");
+        }
         userDao.deleteById(id);
     }
 
@@ -192,8 +211,17 @@ public class UserService {
         Map<String, String> map = new HashMap<>();
         map.put("mobile", mobile);
         map.put("checkCode", checkCode);
-        rabbitTemplate.convertAndSend("sms", map);
+//        rabbitTemplate.convertAndSend("sms", map);
         //在控制台显示一份 (方便测试)
         System.out.println("验证码为:" + checkCode);
+    }
+
+
+    public User login(String mobile, String password) {
+        User user = userDao.findByMobile(mobile);
+        if (user != null && encoder.matches(password, user.getPassword())) {
+            return user;
+        }
+        return null;
     }
 }
